+++ plugin-packs/postcss-preset-env/CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changes to PostCSS Preset Env
 
+### Unreleased (minor)
+
+- Added `@csstools/postcss-position-area-property` [Check the plugin README](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-position-area-property#readme) for usage details.
+- Updated [`cssdb`](https://github.com/csstools/cssdb) to [`8.5.1`](https://github.com/csstools/cssdb/blob/main/CHANGELOG.md#851-december-3-2025) (patch)
+
 ### 10.4.0
 
 _September 21, 2025_
-
-_August 22, 2025_
 
 - Added `@csstools/postcss-contrast-color-function` [Check the plugin README](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-contrast-color-function#readme) for usage details.
 - Updated [`@csstools/postcss-progressive-custom-properties`](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-progressive-custom-properties) to [`4.2.1`](https://github.com/csstools/postcss-plugins/tree/main/plugins/postcss-progressive-custom-properties/CHANGELOG.md#421) (patch)